move LocalImageAsset
