+++ shared/src/androidMain/kotlin/tech/takahana/iconwallpaper/shared/assets/LocalImageAsset.kt
@@ -1,6 +1,8 @@
-package tech.takahana.iconwallpaper.shared.domain.domainobject
+package tech.takahana.iconwallpaper.shared.assets
 
 import androidx.annotation.DrawableRes
+import tech.takahana.iconwallpaper.shared.domain.domainobject.AssetName
+import tech.takahana.iconwallpaper.shared.domain.domainobject.ImageAsset
 
 actual data class LocalImageAsset actual constructor(
     val name: AssetName
