+++ shared/src/commonMain/kotlin/tech/takahana/iconwallpaper/shared/assets/LocalImageAsset.kt
@@ -1,0 +1,8 @@
+package tech.takahana.iconwallpaper.shared.assets
+
+import tech.takahana.iconwallpaper.shared.domain.domainobject.AssetName
+import tech.takahana.iconwallpaper.shared.domain.domainobject.ImageAsset
+
+expect class LocalImageAsset(
+    name: AssetName
+) : ImageAsset
