--- shared/src/commonMain/kotlin/tech/takahana/iconwallpaper/shared/domain/domainobject/LocalImageAsset.kt
@@ -1,5 +1,0 @@
-package tech.takahana.iconwallpaper.shared.domain.domainobject
-
-expect class LocalImageAsset(
-    name: AssetName
-) : ImageAsset
