+++ shared/src/iosMain/kotlin/tech/takahana/iconwallpaper/shared/assets/LocalImageAsset.kt
@@ -1,0 +1,8 @@
+package tech.takahana.iconwallpaper.shared.assets
+
+import tech.takahana.iconwallpaper.shared.domain.domainobject.AssetName
+import tech.takahana.iconwallpaper.shared.domain.domainobject.ImageAsset
+
+actual data class LocalImageAsset actual constructor(
+    val name: AssetName
+) : ImageAsset
