--- shared/src/iosMain/kotlin/tech/takahana/iconwallpaper/shared/domain/domainobject/LocalImageAsset.kt
@@ -1,5 +1,0 @@
-package tech.takahana.iconwallpaper.shared.domain.domainobject
-
-actual data class LocalImageAsset actual constructor(
-    val name: AssetName
-) : ImageAsset
